+++ src/transport/mod.rs
@@ -1,4 +1,4 @@
-#[cfg(all(feature = "ipc-transport", target_family = "unix"))]
+#[cfg(all(feature = "ipc-transport", any(target_family = "unix", windows)))]
 mod ipc;
 #[cfg(feature = "tcp-transport")]
 mod tcp;
@@ -30,7 +30,7 @@ pub(crate) async fn connect(endpoint: &Endpoint) -> ZmqResult<(FramedIo, Endpoin
             do_if_enabled!("tcp-transport", tcp::connect(_host, *_port).await)
         }
         Endpoint::Ipc(_path) => {
-            #[cfg(all(feature = "ipc-transport", target_family = "unix"))]
+            #[cfg(all(feature = "ipc-transport", any(target_family = "unix", windows)))]
             {
                 if let Some(path) = _path {
                     ipc::connect(path).await
@@ -40,7 +40,7 @@ pub(crate) async fn connect(endpoint: &Endpoint) -> ZmqResult<(FramedIo, Endpoin
                     ))
                 }
             }
-            #[cfg(not(all(feature = "ipc-transport", target_family = "unix")))]
+            #[cfg(not(all(feature = "ipc-transport", any(target_family = "unix", windows))))]
             panic!("IPC transport is not available on this platform")
         }
     }
@@ -73,7 +73,7 @@ where
             tcp::begin_accept(_host, _port, _cback).await
         ),
         Endpoint::Ipc(_path) => {
-            #[cfg(all(feature = "ipc-transport", target_family = "unix"))]
+            #[cfg(all(feature = "ipc-transport", any(target_family = "unix", windows)))]
             {
                 if let Some(path) = _path {
                     ipc::begin_accept(&path, _cback).await
@@ -83,7 +83,7 @@ where
                     ))
                 }
             }
-            #[cfg(not(all(feature = "ipc-transport", target_family = "unix")))]
+            #[cfg(not(all(feature = "ipc-transport", any(target_family = "unix", windows))))]
             panic!("IPC transport is not available on this platform")
         }
     }